refactor: move services into app directory
